move term functions to dedicated file
